+++ src/test/java/c/m/impl/LinearFFTest.java
@@ -10,17 +10,17 @@ public class LinearFFTest extends LinearFF implements FitnessFunction{
 
     @Test
     public void mseTest(){
-        Assert.assertEquals(0, result("101", new long[]{1, 2, 3, 4, 5}, new long[]{5, 10, 15, 20, 25}));
+        Assert.assertEquals(0, result("101", new long[]{1, 2, 3, 4, 5}, new double[]{5, 10, 15, 20, 25}));
     }
 
     @Test
     public void mseTest2(){
-        Assert.assertEquals(0, result("101", new long[]{1, 1, 1, 1, 1}, new long[]{5, 5, 5, 5, 5}));
+        Assert.assertEquals(0, result("101", new long[]{1, 1, 1, 1, 1}, new double[]{5, 5, 5, 5, 5}));
     }
 
     @Test
     public void mseTest3(){
-        Assert.assertEquals(2, result("010", new long[]{1, 1, 1, 1, 1}, new long[]{4, 4, 4, 4, 4}));
+        Assert.assertEquals(2, result("010", new long[]{1, 1, 1, 1, 1}, new double[]{4, 4, 4, 4, 4}));
     }
 
 }
